JSON output for ping
Introduce a "-j" argument to have `ping` optionally return machine readable
JSON objects instead of the default text. The output with "-j" can be
piped into JSON parsing tools such as `jq`.
To enable the functionality, a dependency on libjson-c is introduced, however
care was taken for the application to continue operating normally if built
without this dependency - only "-j" will not be available in this case.

TODO: all errors and warnings to JSON
TODO: optional output options (such as verbose) to JSON
^ ideally, _all_ output variations should be valid JSON if combined
with -j

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>